backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
pydantic
pytest
pytest-cov
httpx
